Move LocaleError to a separate module
