+++ rust/agama-server/src/l10n.rs
@@ -1,4 +1,5 @@
 mod dbus;
+pub mod error;
 pub mod helpers;
 mod keyboard;
 pub mod l10n;
@@ -7,6 +8,7 @@ mod timezone;
 pub mod web;
 
 pub use dbus::export_dbus_objects;
+pub use error::LocaleError;
 pub use keyboard::Keymap;
 pub use l10n::L10n;
 pub use locale::LocaleEntry;
